Use 'less -F' for all tty output.
The convoluted logic in PipeToAlternateIfOutputIsLong was all just
trying to re-implement the functionality that turns out to be built in
to 'less -F'.  So just use that instead.

This removes the long_output crate and the atty dependency.
