+++ main/Cargo.toml
@@ -4,7 +4,6 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-atty = "0.2"
 clap = "3.1"
 directories = "4.0"
 term_size = "0.3"
@@ -14,7 +13,6 @@ app = {path = "../app" }
 cli = { path = "../cli" }
 clock = { path = "../clock" }
 config = { path = "../config" }
-long_output = { path = "../long_output" }
 model = { path = "../model" }
 printing = { path = "../printing" }
 text_editing = { path = "../text_editing" }
